+++ index.html
@@ -235,5 +235,5 @@
     </template>
 
 </body>
-<script src="/js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 </html>
