+++ js/setup.js
@@ -1,0 +1,31 @@
+const NAMES = [
+  'Николай',
+  'Екатерина',
+  'Максим',
+  'Никита',
+  'Сергей',
+  'Татьяна',
+  'Андрей',
+  'Людмила'];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTIONS = [
+  'Моё лучшее фото!',
+  'Лето 2023 года',
+  'Первый раз в первый класс',
+  'Когда то это казалось весёлым',
+  'Сколько можно?',
+  'До свидания',
+  'Это был прекрасный день'
+];
+const SOME_USERS = 25;
+
+export { NAMES, MESSAGES, DESCRIPTIONS, SOME_USERS };
